#12 erase and export actions for guests
