+++ Controller/AbstractPrivacy.php
@@ -7,12 +7,6 @@ declare(strict_types=1);
 
 namespace Opengento\Gdpr\Controller;
 
-use Magento\Customer\Controller\AbstractAccount;
-use Magento\Framework\App\Action\Context;
-use Magento\Framework\Controller\ResultFactory;
-use Magento\Framework\Controller\ResultInterface;
-use Opengento\Gdpr\Model\Config;
-
 /**
  * Controller AbstractPrivacy
  */
