+++ ViewModel/Privacy/EraseCustomerDataProvider.php
@@ -7,18 +7,15 @@ declare(strict_types=1);
 
 namespace Opengento\Gdpr\ViewModel\Privacy;
 
-use Magento\Cms\Block\Block;
 use Magento\Customer\Model\Session;
 use Magento\Framework\DataObject;
 use Magento\Framework\View\Element\Block\ArgumentInterface;
-use Magento\Framework\View\Element\BlockFactory;
 use Opengento\Gdpr\Api\EraseCustomerCheckerInterface;
-use Opengento\Gdpr\Model\Config;
 
 /**
- * Class ErasureDataProvider
+ * Class EraseCustomerDataProvider
  */
-final class ErasureDataProvider extends DataObject implements ArgumentInterface
+final class EraseCustomerDataProvider extends DataObject implements ArgumentInterface
 {
     /**
      * @var \Opengento\Gdpr\Api\EraseCustomerCheckerInterface
@@ -26,67 +23,23 @@ final class ErasureDataProvider extends DataObject implements ArgumentInterface
     private $eraseCustomerChecker;
 
     /**
-     * @var \Opengento\Gdpr\Model\Config
-     */
-    private $config;
-
-    /**
      * @var \Magento\Customer\Model\Session
      */
     private $session;
 
     /**
-     * @var \Magento\Framework\View\Element\BlockFactory
-     */
-    private $blockFactory;
-
-    /**
      * @param \Opengento\Gdpr\Api\EraseCustomerCheckerInterface $eraseCustomerChecker
-     * @param \Opengento\Gdpr\Model\Config $config
      * @param \Magento\Customer\Model\Session $session
-     * @param \Magento\Framework\View\Element\BlockFactory $blockFactory
      * @param array $data
      */
     public function __construct(
         EraseCustomerCheckerInterface $eraseCustomerChecker,
-        Config $config,
         Session $session,
-        BlockFactory $blockFactory,
         array $data = []
     ) {
         $this->eraseCustomerChecker = $eraseCustomerChecker;
-        $this->config = $config;
         $this->session = $session;
-        $this->blockFactory = $blockFactory;
         parent::__construct($data);
-    }
-
-    /**
-     * Check if the erasure is enabled
-     *
-     * @return bool
-     */
-    public function isErasureEnabled(): bool
-    {
-        return $this->config->isExportEnabled();
-    }
-
-    /**
-     * Retrieve the erase information html
-     *
-     * @return string
-     */
-    public function getErasureInformation(): string
-    {
-        if (!$this->hasData('erasure_information')) {
-            $block = $this->blockFactory->createBlock(
-                Block::class,
-                ['data' => ['block_id' => $this->config->getErasureInformationBlockId()]]
-            );
-            $this->setData('erasure_information', $block->toHtml());
-        }
-
-        return (string) $this->_getData('erasure_information');
     }
 
     /**
@@ -115,23 +68,5 @@ final class ErasureDataProvider extends DataObject implements ArgumentInterface
         }
 
         return (bool) $this->_getData('can_create');
-    }
-
-    /**
-     * Retrieve the anonymize information html
-     *
-     * @return string
-     */
-    public function getAnonymizeInformation(): string
-    {
-        if (!$this->hasData('anonymize_information')) {
-            $block = $this->blockFactory->createBlock(
-                Block::class,
-                ['data' => ['block_id' => $this->config->getAnonymizeInformationBlockId()]]
-            );
-            $this->setData('anonymize_information', $block->toHtml());
-        }
-
-        return (string) $this->_getData('anonymize_information');
     }
 }
