+++ view/frontend/layout/customer_privacy_settings.xml
@@ -25,7 +25,7 @@
                 </block>
                 <block class="Magento\Framework\View\Element\Template" name="customer-account-privacy-erasure" as="erasure" template="Opengento_Gdpr::account/privacy/erase.phtml" ifconfig="gdpr/erasure/enabled" cacheable="false">
                     <arguments>
-                        <argument name="erasureDataProvider" xsi:type="object">Opengento\Gdpr\ViewModel\Privacy\ErasureDataProvider</argument>
+                        <argument name="eraseDataProvider" xsi:type="object">Opengento\Gdpr\ViewModel\Privacy\EraseDataProvider</argument>
                         <argument name="urlsDataProvider" xsi:type="object">Opengento\Gdpr\ViewModel\UrlsDataProvider</argument>
                     </arguments>
                 </block>
