+++ view/frontend/layout/default.xml
@@ -8,7 +8,7 @@
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <body>
         <referenceContainer name="before.body.end">
-            <block class="Opengento\Gdpr\Block\Messages\PrivacyMessagePopup" name="privacy_message_popup" after="-" ifconfig="gdpr/general/enabled">
+            <block class="Opengento\Gdpr\Block\Messages\PrivacyMessagePopup" name="privacy_message_popup" after="-" ifconfig="gdpr/cookie/enabled">
                 <arguments>
                     <argument name="template" xsi:type="string">Opengento_Gdpr::messages/popup.phtml</argument>
                     <argument name="jsLayout" xsi:type="array">
