+++ view/frontend/layout/sales_guest_view.xml
@@ -8,11 +8,29 @@
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <body>
         <referenceBlock name="sales.order.info.buttons">
-            <block class="Magento\Framework\View\Element\Template" as="buttons.export" name="sales.order.info.buttons.export" template="Opengento_Gdpr::order/info/buttons/export.phtml" cacheable="false">
+            <block class="Magento\Framework\View\Element\Template" name="sales.order.info.buttons.export" as="buttons.export" template="Opengento_Gdpr::order/info/buttons/export.phtml" ifconfig="gdpr/export/enabled">
                 <arguments>
                     <argument name="actionUrl" xsi:type="url" path="customer/guest/export"/>
                 </arguments>
             </block>
+            <block class="Magento\Framework\View\Element\Template" name="sales.order.info.buttons.erase" as="buttons.erase" template="Opengento_Gdpr::order/info/buttons/erase.phtml" ifconfig="gdpr/erasure/enabled">
+                <arguments>
+                    <argument name="actionUrl" xsi:type="url" path="customer/guest/erase"/>
+                    <argument name="eraseGuestDataProvider" xsi:type="object">Opengento\Gdpr\ViewModel\Privacy\EraseGuestDataProvider</argument>
+                </arguments>
+            </block>
         </referenceBlock>
+        <referenceContainer name="content">
+            <block class="Magento\Framework\View\Element\Template" name="sales.order.info.export" template="Opengento_Gdpr::order/info/export.phtml" ifconfig="gdpr/export/enabled">
+                <arguments>
+                    <argument name="exportDataProvider" xsi:type="object">Opengento\Gdpr\ViewModel\Privacy\ExportDataProvider</argument>
+                </arguments>
+            </block>
+            <block class="Magento\Framework\View\Element\Template" name="sales.order.info.erase" template="Opengento_Gdpr::order/info/erase.phtml" ifconfig="gdpr/erasure/enabled">
+                <arguments>
+                    <argument name="eraseDataProvider" xsi:type="object">Opengento\Gdpr\ViewModel\Privacy\EraseDataProvider</argument>
+                </arguments>
+            </block>
+        </referenceContainer>
     </body>
 </page>
