+++ view/frontend/templates/order/info/erase.phtml
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Copyright © OpenGento, All rights reserved.
+ * See LICENSE bundled with this library for license details.
+ */
+
+/** @var \Magento\Framework\View\Element\Template $block */
+
+/** @var \Opengento\Gdpr\ViewModel\Privacy\EraseDataProvider $eraseDataProvider */
+$eraseDataProvider = $block->getData('eraseDataProvider');
+?>
+<div class="block block-order-details-view">
+    <div class="block-title">
+        <strong><?= $block->escapeHtml(__('Right to be forgotten')) ?></strong>
+    </div>
+    <div class="block-content">
+        <div class="widget block block-static-block">
+            <?= $eraseDataProvider->getErasureInformation() ?>
+        </div>
+    </div>
+</div>
