+++ view/frontend/web/css/source/_module.less
@@ -1,4 +1,29 @@
 /**
+ * Copyright © OpenGento, All rights reserved.
+ * See LICENSE bundled with this library for license details.
+ */
+
+@color-red: #ff0101;
+@color-red2: #b30000;
+@color-gray-light01: #f4f4f4;
+@screen__m: 768px;
+
+&.action {
+  &.action-danger {
+    color: @color-red9;
+  }
+  &.primary.primary-danger {
+    background: @color-red9;
+    border: 1px solid @color-red9;
+
+    &:hover {
+      background: @color-red11;
+      border: 1px solid @color-red11;
+    }
+  }
+}
+
+/**
  * This file is part of the Flurrybox EnhancedPrivacy package.
  *
  * DISCLAIMER
@@ -12,21 +37,6 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
-@color-red: #ff0101;
-@color-red2: #b30000;
-@color-gray-light01: #f4f4f4;
-@screen__m: 768px;
-
-&.action.primary.primary-danger {
-  background: @color-red9;
-  border: 1px solid @color-red9;
-
-  &:hover {
-    background: @color-red11;
-    border: 1px solid @color-red11;
-  }
-}
 
 .enhanced-privacy-popup {
   position: absolute;
